+++ src/components/subway/SubwayList.tsx
@@ -2,9 +2,7 @@ import { SubwayLocationObj } from '@/types/subway';
 import { InfoContainer } from '@/components/maps/infos';
 import tw, { styled } from 'twin.macro';
 import SubwayName from './SubwayName';
-import Button from '../common/Button/Button';
-import { useRecoilState } from 'recoil';
-import { dayState } from '@/recoil/atoms/dayState';
+import Days from './Days';
 
 const Container = styled.section([
   tw`mobile:(w-full h-auto)`,
@@ -22,35 +20,11 @@ const SubwayList = ({
   line: number;
   subwayList: SubwayLocationObj[] | null;
 }) => {
-  const [days, setDays] = useRecoilState(dayState);
-  const handleDays = (e: React.MouseEvent<HTMLButtonElement>) => {
-    if (!e.currentTarget.textContent) return;
-
-    setDays(e.currentTarget.textContent);
-  };
-
   return (
     <Container>
       <div css={tw`flex justify-between items-center w-full`}>
         <SubwayName />
-        <div tw='inline-flex' role='group'>
-          <Button
-            onClick={handleDays}
-            type={days === '평일' ? 'primary' : 'default'}
-            size='small'
-            props={[tw`text-14 mb-20`]}
-          >
-            평일
-          </Button>
-          <Button
-            onClick={handleDays}
-            type={days === '토요일' ? 'primary' : 'default'}
-            size='small'
-            props={[tw`text-14 mb-20`]}
-          >
-            토요일
-          </Button>
-        </div>
+        <Days />
       </div>
       <div
         css={[
